Build fixes for the Percona 8.4.10 merge (inherited, macOS/portability)
None of these are merge regressions. Four separate breakages that block the
build on macOS/arm64; all verified against the parents before changing anything.

components/audit_log_filter/audit_udf.cc, .../log_writer/file_handle.cc and the
two sprintf sites in sql/mysqld.cc are byte-identical to Percona (git diff
against Percona-Server-8.4.10-10 is empty for the first two; for mysqld.cc,
which we did resolve conflicts in, the offending lines appear verbatim in
Percona at :11836/:11846 and origin/main contains zero occurrences of the
format string).

  Format-string mismatches: int64_t/uint64_t are 'long' on LP64 Linux but
  'long long' on macOS/arm64, so %ld/%lu trip -Werror,-Wformat. Percona's CI is
  Linux-only, so these never fire for them. Fixed portably with a cast to the
  widest type rather than a platform #ifdef.

  file_handle.cc: flush_on_write is only consumed inside #ifdef O_DIRECT.
  macOS has no O_DIRECT, so the #else branch compiles and the parameter is
  genuinely unused -> -Werror,-Wunused-parameter. Marked [[maybe_unused]],
  matching the style already used elsewhere in that component.

  unittest/gunit/CMakeLists.txt: a *configure*-time error, not a compile error.
  sqlgunitlib is linked with the keyword signature on one line and Percona's
  added line used the plain signature; CMake forbids mixing on one target. The
  file auto-merged cleanly -- the clash was semantic. The 8.4.8 merge hit this
  identical failure.

Worth reporting the format-string ones upstream to Percona; if they take them,
our divergence in that component goes to zero at the next merge.

+++ sql/mysqld.cc
@@ -11928,8 +11928,11 @@ static int show_threadpool_average_queue_wait_us(THD *thd [[maybe_unused]],
   var->type = SHOW_CHAR;
   var->value = buff;
   auto stat = tp_get_average_queue_wait_stats();
-  sprintf(buff, "avg: %.3f, min: %.3f, max: %.3f, dev: %.3f, cnt: %ld",
-          stat.average, stat.min, stat.max, stat.deviation, stat.count);
+  // stat.count is int64_t, which is 'long' on LP64 Linux but 'long long' on
+  // macOS/arm64, so a bare %ld fails -Werror=format there.
+  sprintf(buff, "avg: %.3f, min: %.3f, max: %.3f, dev: %.3f, cnt: %lld",
+          stat.average, stat.min, stat.max, stat.deviation,
+          static_cast<long long>(stat.count));
   return 0;
 }
 
@@ -11938,8 +11941,11 @@ static int show_threadpool_average_hp_queue_wait_us(THD *thd [[maybe_unused]],
   var->type = SHOW_CHAR;
   var->value = buff;
   auto stat = tp_get_average_hp_queue_wait_stats();
-  sprintf(buff, "avg: %.3f, min: %.3f, max: %.3f, dev: %.3f, cnt: %ld",
-          stat.average, stat.min, stat.max, stat.deviation, stat.count);
+  // stat.count is int64_t, which is 'long' on LP64 Linux but 'long long' on
+  // macOS/arm64, so a bare %ld fails -Werror=format there.
+  sprintf(buff, "avg: %.3f, min: %.3f, max: %.3f, dev: %.3f, cnt: %lld",
+          stat.average, stat.min, stat.max, stat.deviation,
+          static_cast<long long>(stat.count));
   return 0;
 }
